Simple numeric value parsing and output to screen
The text editor relies on several things to correlate input to a glyph
on a bmp font. First is the initial x, y co-ordinates of the
alphanumeric series, such as where the 'a' letter is on the font sheet
or where '0' is for a numeric series.

We then have the length of the series (e.g. 26 chars in the alphabet, 10
chars for 0-9). After that we take user input and map it to a index in
the specified series which we use to specify the rect on the font sheet
to render.