Refactor canonicaliseCaretPos, change caret pos
Moving the caret at top of document no longer snaps the caret to the
start of the document but remains in place.

Caret positioning is now represented as a 1D value, which the program
manipulates. The caret's position is converted to x, y coordinates only
when needed. This reduces the LOC needed to do bounds checking.