Game instantiates objs from reference when needed
There's a global list of objects stored in memory of all possible game
objects. The game instantiates from this list on demand. Large objects
may be disposed from the list after creation if needed and replaced
with a null entry.

Enemy now drops coins according to how many it has in possession.