Generate archive name based off files to back up
There are several cases. If a singular file, the file can be a directory,
absolute path or relative path. The file name will be used as the archive
name. If it's a path, extract the filename from the path.

If there are more than one file as part of the backup, then we use the
current directory the file resides in as the archive name.