Add path checking, append '\' to path if needed
Add the '\' so Windows treats it as a path, not file. The WINAPI has this
function under PathCchAddBackslash(), it is another library to include.
I have chosen not to add this for now since it requires a pointer to
a path. We know our paths can't be greater than MAX_PATH(260) so our paths
are currently defined as path[MAX_PATH].

Our own implementation is rather trivial without having to generate
a pointer to a string so it has been left as such.