Add hard-linking of files if possible else copy
Hard-linking is the default behaviour until CFG token is created for it.
Hard-link saves space whereby each copy of the file is a link to the
original. File is not deleted until all links are deleted. The downside is
corruption in the original backup invalidates the file.

Hard-linking fails if the file is on a different drive in which the
fallsback to copying the file instead.