Add debug service for OSD variable stack, change audio handling
Small changes to audio handling to collate all media state changes into a single
function for ease of reading and ensuring the right calls are made when sending
state changes to the media player minimising crashes from invalid states.

Additional to that prompted debug services to display variable state on screen
to ease tracking of state and generally help manage the application. Minimise
the number of times required to invoke a debugger, or building through Android
Studio because it's so slow.