Fix skip next/prev button not working on startup
The next/prev buttons only worked when a playlist has been bound to the audio
service. On startup, the active playlist for the main activity is set, and
attempted to bind to the service. Since the service has playlist validation, the
playlist appeared to be bound in the app on the client but not the service since
the index was not initialised properly (-1, because we can't guarantee that we
have any elements in the playlist, such as an empty playlist).

Instead defer the binding until data has been loaded during the device query,
then display.