Fix search mode overriding active playlist
In search mode, upon choosing a song the active playlist will switch to the
search playlist which is non-accessible by the user in the UI as it does not
show up in the playlist list. Hence, re-activating search mode will being
replace the one of the active playlists with the search permanently.