AudioService uses playspec ref, add debug checkbox
Remove doubling of state in AudioService and MainActivity, i.e. AudioService now
has access to all playback data to reduce the need to have to ensure updates are
propagated to the service by having access to all playback state.

Add checkbox marker to clearly identify when debug overlay is set.