Refine single threaded approach to syncing
Monitoring handles time out such that we can periodically check if there are
unsaved changes to be backed up making it more reliable. Currently the
granularity is set at 500ms, half a second. With the downside of the tray icon
being half a second slow, and the app is continuously running in background.

But it does now stop the thread from being blocked and not responding to
messages the window receives.